Moved fetching data to RX

+++ app/src/main/java/com/stas/tsepa/sandboxapp/cloud/ObservableLecturesAPI.java
@@ -10,5 +10,5 @@ import rx.Observable;
 
 public interface ObservableLecturesAPI {
     @GET("lecture?token=E3yK8syiWdH6du&expand=courses")
-    Observable<List<LectureItem>> loadItems(@Query("page") int page, @Query("per-page") int perPage);
+    Observable<List<LectureItem>> loadItems(@Query("page") long page, @Query("per-page") int perPage);
 }
